# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Eternal Kingdoms API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kingdoms/mine:
    get:
      operationId: getMyKingdom
      tags: [kingdoms]
      summary: Get the authenticated user's kingdom
      responses:
        "200":
          description: User's kingdom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kingdom"
        "404":
          description: Kingdom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kingdoms/{id}:
    get:
      operationId: getKingdom
      tags: [kingdoms]
      summary: Get a kingdom by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kingdom details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kingdom"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kingdoms/{id}/summary:
    get:
      operationId: getKingdomSummary
      tags: [kingdoms]
      summary: Get a kingdom's dashboard summary (resources, buildings, troops)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kingdom summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KingdomSummary"
  /kingdoms/{id}/buildings:
    get:
      operationId: getKingdomBuildings
      tags: [buildings]
      summary: Get all buildings in a kingdom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
  /kingdoms/{id}/resources:
    get:
      operationId: getKingdomResources
      tags: [resources]
      summary: Get current resources for a kingdom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kingdom resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resources"
...
  /kingdoms/{id}/construction-queue:
    get:
      operationId: getConstructionQueue
      tags: [buildings]
      summary: Get active new-building construction queue for a kingdom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active construction queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConstructionQueueItem"
        "404":
          description: Kingdom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kingdoms/{id}/construction-options:
    get:
      operationId: getConstructionOptions
      tags: [buildings]
      summary: Get available building types to construct with costs and slot availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Construction options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionOptions"
        "404":
          description: Kingdom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kingdoms/{id}/construct:
    post:
      operationId: constructBuilding
      tags: [buildings]
      summary: Start construction of a new building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructBuildingInput"
      responses:
        "202":
          description: Construction started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionResponse"
        "402":
          description: Insufficient resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientResourcesError"
        "409":
          description: Slot limit reached or palace tier insufficient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kingdom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /heroes:
    get:
      operationId: listHeroes
      tags: [heroes]
      summary: List heroes for a kingdom
      parameters:
        - name: kingdomId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hero list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroListResponse"
  /heroes/{id}:
    get:
      operationId: getHero
      tags: [heroes]
      summary: Get a single hero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hero detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroResponse"

  # ── Phase 4: Troops ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Kingdom:
      type: object
      required: [id, userId, name, worldId, power, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        worldId:
          type: integer
        mapX:
          type: ["integer", "null"]
        mapY:
          type: ["integer", "null"]
        power:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    KingdomSummary:
      type: object
      required: [kingdom, resources, buildingCount, troopCount]
      properties:
        kingdom:
          $ref: "#/components/schemas/Kingdom"
        resources:
          $ref: "#/components/schemas/Resources"
        buildingCount:
          type: integer
        troopCount:
          type: integer
        power:
          type: integer
    Building:
      type: object
      required: [id, kingdomId, buildingType, level, assetId]
      properties:
        id:
          type: integer
        kingdomId:
          type: integer
        buildingType:
          type: string
          enum: [palace, farm, lumber_mill, quarry, iron_mine, gold_mine, barracks, wall, watch_tower]
        level:
          type: integer
        assetId:
          type: string
        isConstructing:
          type: boolean
        constructionEndsAt:
          type: ["string", "null"]
          format: date-time
        positionX:
          type: ["integer", "null"]
        positionY:
          type: ["integer", "null"]
    Resources:
      type: object
      required: [kingdomId, food, wood, stone, iron, gold, updatedAt]
      properties:
        kingdomId:
          type: integer
        food:
          type: integer
        wood:
          type: integer
        stone:
          type: integer
        iron:
          type: integer
        gold:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    ResourceAmounts:
      type: object
      required: [food, wood, stone, iron, gold]
      properties:
        food:
          type: integer
        wood:
          type: integer
        stone:
          type: integer
        iron:
          type: integer
        gold:
          type: integer
    PalaceTierConfig:
      type: object
      required: [maxResourceBuildings, maxBarracks, maxWalls, maxWatchTowers, unlocksFeatures]
      properties:
        maxResourceBuildings:
          type: integer
        maxBarracks:
          type: integer
        maxWalls:
          type: integer
        maxWatchTowers:
          type: integer
        unlocksFeatures:
          type: array
          items:
            type: string
...
    InsufficientResourcesError:
      type: object
      required: [error, required]
      properties:
        error:
          type: string
        shortfall:
          $ref: "#/components/schemas/ResourceAmounts"
        required:
          $ref: "#/components/schemas/ResourceAmounts"
...
    ConstructionQueueItem:
      type: object
      required: [id, kingdomId, buildingId, buildingType, startsAt, endsAt, status]
      properties:
        id:
          type: integer
        kingdomId:
          type: integer
        buildingId:
          type: integer
        buildingType:
          type: string
          enum: [palace, farm, lumber_mill, quarry, iron_mine, gold_mine, barracks, wall, watch_tower]
        foodCost:
          type: integer
        woodCost:
          type: integer
        stoneCost:
          type: integer
        ironCost:
          type: integer
        goldCost:
          type: integer
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [in_progress, completed, cancelled]
    ConstructionOption:
      type: object
      required: [buildingType, label, cost, durationSeconds, slotAvailable, canAfford, shortfall]
      properties:
        buildingType:
          type: string
          enum: [palace, farm, lumber_mill, quarry, iron_mine, gold_mine, barracks, wall, watch_tower]
        label:
          type: string
        cost:
          $ref: "#/components/schemas/ResourceAmounts"
        durationSeconds:
          type: integer
        slotAvailable:
          type: boolean
        slotReason:
          type: ["string", "null"]
        canAfford:
          type: boolean
        shortfall:
          $ref: "#/components/schemas/ResourceAmounts"
    ConstructionOptions:
      type: object
      required: [palaceLevel, palaceTier, options]
      properties:
        palaceLevel:
          type: integer
        palaceTier:
          $ref: "#/components/schemas/PalaceTierConfig"
        options:
          type: array
          items:
            $ref: "#/components/schemas/ConstructionOption"
    ConstructBuildingInput:
      type: object
      required: [buildingType]
      properties:
        buildingType:
          type: string
          enum: [farm, lumber_mill, quarry, iron_mine, gold_mine, barracks, wall, watch_tower]
    ConstructionResponse:
      type: object
      required: [message, construction, building, durationSeconds, endsAt, costDeducted]
      properties:
        message:
          type: string
        construction:
          $ref: "#/components/schemas/ConstructionQueueItem"
        building:
          $ref: "#/components/schemas/Building"
        durationSeconds:
          type: integer
        endsAt:
          type: string
          format: date-time
        costDeducted:
          $ref: "#/components/schemas/ResourceAmounts"
...
    HeroStats:
      type: object
      required: [command, attack, defense, speed, gathering]
      properties:
        command:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        speed:
          type: integer
        gathering:
          type: integer
    HeroSkill:
      type: object
      required: [skillId, name, description, effectType, effectValue]
      properties:
        skillId:
          type: string
        name:
          type: string
        description:
          type: string
        triggerRound:
          type: integer
        effectType:
          type: string
        effectValue:
          type: number
    HeroDto:
      type: object
      required: [id, kingdomId, assetId, name, rarity, level, experience, experienceToNext, leadershipCapacity, stats, skills, isLeading, createdAt, updatedAt]
      properties:
        id:
          type: integer
        kingdomId:
          type: integer
        assetId:
          type: string
        name:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        level:
          type: integer
        experience:
          type: integer
        experienceToNext:
          type: integer
        leadershipCapacity:
          type: integer
        troopAffinity:
          type: string
        stats:              { $ref: "#/components/schemas/HeroStats" }
        skills:
          type: array
          items: { $ref: "#/components/schemas/HeroSkill" }
        isLeading:
          type: boolean
        nftContractAddress:
          type: string
        nftTokenId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HeroListResponse:
      type: object
      required: [heroes]
      properties:
        heroes:
          type: array
          items: { $ref: "#/components/schemas/HeroDto" }
    HeroResponse:
      type: object
      required: [hero]
      properties:
        hero: { $ref: "#/components/schemas/HeroDto" }
...