# lib/api-zod/src/generated/types/insufficientResourcesError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eternal Kingdoms API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceAmounts } from './resourceAmounts';

export interface InsufficientResourcesError {
  error: string;
  shortfall?: ResourceAmounts;
  required: ResourceAmounts;
}